split QgsPostgresSharedData and QgsPostgresUtils to separate file, to allow their use in qgspostgresrasterprovider
