+++ src/providers/postgres/qgspostgresprovider.cpp
@@ -49,6 +49,7 @@
 #include "qgsprovidermetadata.h"
 #include "qgspostgresproviderconnection.h"
 #include "qgspostgresprovidermetadatautils.h"
+#include "qgspostgresutils.h"
 #include <QRegularExpression>
 
 const QString QgsPostgresProvider::POSTGRES_KEY = QStringLiteral( "postgres" );
@@ -637,193 +638,6 @@ QString QgsPostgresProvider::whereClause( QgsFeatureId featureId ) const
 QString QgsPostgresProvider::whereClause( QgsFeatureIds featureIds ) const
 {
   return QgsPostgresUtils::whereClause( featureIds, mAttributeFields, connectionRO(), mPrimaryKeyType, mPrimaryKeyAttrs, mShared );
-}
-
-
-QString QgsPostgresUtils::whereClause( QgsFeatureId featureId, const QgsFields &fields, QgsPostgresConn *conn, QgsPostgresPrimaryKeyType pkType, const QList<int> &pkAttrs, const std::shared_ptr<QgsPostgresSharedData> &sharedData )
-{
-  QString whereClause;
-
-  switch ( pkType )
-  {
-    case PktTid:
-      whereClause = QStringLiteral( "ctid='(%1,%2)'" )
-                      .arg( FID_TO_NUMBER( featureId ) >> 16 )
-                      .arg( FID_TO_NUMBER( featureId ) & 0xffff );
-      break;
-
-    case PktOid:
-      whereClause = QStringLiteral( "oid=%1" ).arg( featureId );
-      break;
-
-    case PktInt:
-      Q_ASSERT( pkAttrs.size() == 1 );
-      whereClause = QStringLiteral( "%1=%2" ).arg( QgsPostgresConn::quotedIdentifier( fields.at( pkAttrs[0] ).name() ) ).arg( FID2PKINT( featureId ) );
-      break;
-
-    case PktInt64:
-    case PktUint64:
-    {
-      Q_ASSERT( pkAttrs.size() == 1 );
-      QVariantList pkVals = sharedData->lookupKey( featureId );
-      if ( !pkVals.isEmpty() )
-      {
-        QgsField fld = fields.at( pkAttrs[0] );
-        whereClause = conn->fieldExpression( fld );
-        if ( !QgsVariantUtils::isNull( pkVals[0] ) )
-          whereClause += '=' + pkVals[0].toString();
-        else
-          whereClause += QLatin1String( " IS NULL" );
-      }
-    }
-    break;
-
-    case PktFidMap:
-    {
-      QVariantList pkVals = sharedData->lookupKey( featureId );
-      if ( !pkVals.isEmpty() )
-      {
-        Q_ASSERT( pkVals.size() == pkAttrs.size() );
-
-        QString delim;
-        for ( int i = 0; i < pkAttrs.size(); i++ )
-        {
-          int idx = pkAttrs[i];
-          QgsField fld = fields.at( idx );
-
-          whereClause += delim + conn->fieldExpressionForWhereClause( fld, static_cast<QMetaType::Type>( pkVals[i].userType() ) );
-          if ( QgsVariantUtils::isNull( pkVals[i] ) )
-            whereClause += QLatin1String( " IS NULL" );
-          else
-            whereClause += '=' + QgsPostgresConn::quotedValue( pkVals[i] ); // remove toString as it must be handled by quotedValue function
-
-          delim = QStringLiteral( " AND " );
-        }
-      }
-      else
-      {
-        QgsDebugError( QStringLiteral( "FAILURE: Key values for feature %1 not found." ).arg( featureId ) );
-        whereClause = QStringLiteral( "NULL" );
-      }
-    }
-    break;
-
-    case PktUnknown:
-      Q_ASSERT( !"FAILURE: Primary key unknown" );
-      whereClause = QStringLiteral( "NULL" );
-      break;
-  }
-
-  return whereClause;
-}
-
-QString QgsPostgresUtils::whereClause( const QgsFeatureIds &featureIds, const QgsFields &fields, QgsPostgresConn *conn, QgsPostgresPrimaryKeyType pkType, const QList<int> &pkAttrs, const std::shared_ptr<QgsPostgresSharedData> &sharedData )
-{
-  auto lookupKeyWhereClause = [=] {
-    if ( featureIds.isEmpty() )
-      return QString();
-
-    //simple primary key, so prefer to use an "IN (...)" query. These are much faster then multiple chained ...OR... clauses
-    QString delim;
-    QString expr = QStringLiteral( "%1 IN (" ).arg( QgsPostgresConn::quotedIdentifier( fields.at( pkAttrs[0] ).name() ) );
-
-    for ( const QgsFeatureId featureId : std::as_const( featureIds ) )
-    {
-      const QVariantList pkVals = sharedData->lookupKey( featureId );
-      if ( !pkVals.isEmpty() )
-      {
-        expr += delim + QgsPostgresConn::quotedValue( pkVals.at( 0 ) );
-        delim = ',';
-      }
-    }
-    expr += ')';
-
-    return expr;
-  };
-
-  switch ( pkType )
-  {
-    case PktOid:
-    case PktInt:
-    {
-      QString expr;
-
-      //simple primary key, so prefer to use an "IN (...)" query. These are much faster then multiple chained ...OR... clauses
-      if ( !featureIds.isEmpty() )
-      {
-        QString delim;
-        expr = QStringLiteral( "%1 IN (" ).arg( ( pkType == PktOid ? QStringLiteral( "oid" ) : QgsPostgresConn::quotedIdentifier( fields.at( pkAttrs[0] ).name() ) ) );
-
-        for ( const QgsFeatureId featureId : std::as_const( featureIds ) )
-        {
-          expr += delim + FID_TO_STRING( ( pkType == PktOid ? featureId : FID2PKINT( featureId ) ) );
-          delim = ',';
-        }
-        expr += ')';
-      }
-
-      return expr;
-    }
-    case PktInt64:
-    case PktUint64:
-      return lookupKeyWhereClause();
-
-    case PktFidMap:
-    case PktTid:
-    case PktUnknown:
-    {
-      // on simple string primary key we can use IN
-      if ( pkType == PktFidMap && pkAttrs.count() == 1 && fields.at( pkAttrs[0] ).type() == QMetaType::Type::QString )
-        return lookupKeyWhereClause();
-
-      //complex primary key, need to build up where string
-      QStringList whereClauses;
-      for ( const QgsFeatureId featureId : std::as_const( featureIds ) )
-      {
-        whereClauses << whereClause( featureId, fields, conn, pkType, pkAttrs, sharedData );
-      }
-      return whereClauses.isEmpty() ? QString() : whereClauses.join( QLatin1String( " OR " ) ).prepend( '(' ).append( ')' );
-    }
-  }
-  return QString(); //avoid warning
-}
-
-QString QgsPostgresUtils::andWhereClauses( const QString &c1, const QString &c2 )
-{
-  if ( c1.isEmpty() )
-    return c2;
-  if ( c2.isEmpty() )
-    return c1;
-
-  return QStringLiteral( "(%1) AND (%2)" ).arg( c1, c2 );
-}
-
-void QgsPostgresUtils::replaceInvalidXmlChars( QString &xml )
-{
-  static const QRegularExpression replaceRe { QStringLiteral( "([\\x00-\\x08\\x0B-\\x1F\\x7F])" ) };
-  QRegularExpressionMatchIterator it { replaceRe.globalMatch( xml ) };
-  while ( it.hasNext() )
-  {
-    const QRegularExpressionMatch match { it.next() };
-    const QChar c { match.captured( 1 ).at( 0 ) };
-    xml.replace( c, QStringLiteral( "UTF-8[%1]" ).arg( c.unicode() ) );
-  }
-}
-
-void QgsPostgresUtils::restoreInvalidXmlChars( QString &xml )
-{
-  static const QRegularExpression replaceRe { QStringLiteral( R"raw(UTF-8\[(\d+)\])raw" ) };
-  QRegularExpressionMatchIterator it { replaceRe.globalMatch( xml ) };
-  while ( it.hasNext() )
-  {
-    const QRegularExpressionMatch match { it.next() };
-    bool ok;
-    const ushort code { match.captured( 1 ).toUShort( &ok ) };
-    if ( ok )
-    {
-      xml.replace( QStringLiteral( "UTF-8[%1]" ).arg( code ), QChar( code ) );
-    }
-  }
 }
 
 QString QgsPostgresProvider::filterWhereClause() const
@@ -5854,124 +5668,6 @@ void QgsPostgresProviderMetadata::cleanupProvider()
 }
 
 
-// ----------
-
-void QgsPostgresSharedData::addFeaturesCounted( long long diff )
-{
-  QMutexLocker locker( &mMutex );
-
-  if ( mFeaturesCounted >= 0 )
-    mFeaturesCounted += diff;
-}
-
-void QgsPostgresSharedData::ensureFeaturesCountedAtLeast( long long fetched )
-{
-  QMutexLocker locker( &mMutex );
-
-  /* only updates the feature count if it was already once.
-   * Otherwise, this would lead to false feature count if
-   * an existing project is open at a restrictive extent.
-   */
-  if ( mFeaturesCounted > 0 && mFeaturesCounted < fetched )
-  {
-    QgsDebugMsgLevel( QStringLiteral( "feature count adjusted from %1 to %2" ).arg( mFeaturesCounted ).arg( fetched ), 2 );
-    mFeaturesCounted = fetched;
-  }
-}
-
-long long QgsPostgresSharedData::featuresCounted()
-{
-  QMutexLocker locker( &mMutex );
-  return mFeaturesCounted;
-}
-
-void QgsPostgresSharedData::setFeaturesCounted( long long count )
-{
-  QMutexLocker locker( &mMutex );
-  mFeaturesCounted = count;
-}
-
-
-QgsFeatureId QgsPostgresSharedData::lookupFid( const QVariantList &v )
-{
-  QMutexLocker locker( &mMutex );
-
-  QMap<QVariantList, QgsFeatureId>::const_iterator it = mKeyToFid.constFind( v );
-
-  if ( it != mKeyToFid.constEnd() )
-  {
-    return it.value();
-  }
-
-  mFidToKey.insert( ++mFidCounter, v );
-  mKeyToFid.insert( v, mFidCounter );
-
-  return mFidCounter;
-}
-
-
-QVariantList QgsPostgresSharedData::removeFid( QgsFeatureId fid )
-{
-  QMutexLocker locker( &mMutex );
-
-  QVariantList v = mFidToKey[fid];
-  mFidToKey.remove( fid );
-  mKeyToFid.remove( v );
-  return v;
-}
-
-void QgsPostgresSharedData::insertFid( QgsFeatureId fid, const QVariantList &k )
-{
-  QMutexLocker locker( &mMutex );
-
-  mFidToKey.insert( fid, k );
-  mKeyToFid.insert( k, fid );
-}
-
-QVariantList QgsPostgresSharedData::lookupKey( QgsFeatureId featureId )
-{
-  QMutexLocker locker( &mMutex );
-
-  QMap<QgsFeatureId, QVariantList>::const_iterator it = mFidToKey.constFind( featureId );
-  if ( it != mFidToKey.constEnd() )
-    return it.value();
-  return QVariantList();
-}
-
-void QgsPostgresSharedData::clear()
-{
-  QMutexLocker locker( &mMutex );
-  mFidToKey.clear();
-  mKeyToFid.clear();
-  mFeaturesCounted = -1;
-  mFidCounter = 0;
-}
-
-void QgsPostgresSharedData::clearSupportsEnumValuesCache()
-{
-  QMutexLocker locker( &mMutex );
-  mFieldSupportsEnumValues.clear();
-}
-
-bool QgsPostgresSharedData::fieldSupportsEnumValuesIsSet( int index )
-{
-  QMutexLocker locker( &mMutex );
-  return mFieldSupportsEnumValues.contains( index );
-}
-
-bool QgsPostgresSharedData::fieldSupportsEnumValues( int index )
-{
-  QMutexLocker locker( &mMutex );
-  return mFieldSupportsEnumValues.contains( index ) && mFieldSupportsEnumValues[index];
-}
-
-void QgsPostgresSharedData::setFieldSupportsEnumValues( int index, bool isSupported )
-{
-  QMutexLocker locker( &mMutex );
-  mFieldSupportsEnumValues[index] = isSupported;
-}
-
-
 QgsPostgresProviderMetadata::QgsPostgresProviderMetadata()
   : QgsProviderMetadata( QgsPostgresProvider::POSTGRES_KEY, QgsPostgresProvider::POSTGRES_DESCRIPTION )
 {
@@ -6037,7 +5733,6 @@ QVariantMap QgsPostgresProviderMetadata::decodeUri( const QString &uri ) const
 
   return uriParts;
 }
-
 
 QString QgsPostgresProviderMetadata::encodeUri( const QVariantMap &parts ) const
 {
